Amélioration gestion d'erreur Collaborateur

+++ back/vroum_vroum/src/main/java/demo/vroum_vroum/services/CollaborateurService.java
@@ -1,8 +1,10 @@
 package demo.vroum_vroum.services;
 
 import demo.vroum_vroum.entities.Collaborateur;
+import demo.vroum_vroum.exceptions.NotAuthenticatedException;
 import demo.vroum_vroum.repositories.CollaborateurRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import jakarta.persistence.EntityNotFoundException;
+
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
@@ -16,7 +18,11 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
+/**
+ * Service lié à l'entité Collaborateur
+ */
 @Service
 public class CollaborateurService implements UserDetailsService {
     /** Repository de l'entité Collaborateur */
@@ -26,32 +32,54 @@ public class CollaborateurService implements UserDetailsService {
      * Constructeur
      * @param collaborateurRepository repo Collaborateur
      */
-    @Autowired
     public CollaborateurService(CollaborateurRepository collaborateurRepository) {
         this.collaborateurRepository = collaborateurRepository;
     }
 
-
+    /**
+     * Récupère l'utilisateur à partir de son pseudo.
+     * @param pseudo pseudo utilisateur
+     * @return l'optional d'une entité Collaborateur
+     */
     public Collaborateur findByPseudo(String pseudo) {
-        return collaborateurRepository.findByPseudo(pseudo);
+        Optional<Collaborateur> collaborateur = collaborateurRepository.findByPseudo(pseudo);
+
+        if (collaborateur.isEmpty()) {
+            throw new EntityNotFoundException("Pas d'utilisateur trouvé avec le pseudo " + pseudo);
+        }
+
+        return collaborateur.get();
     }
 
-    public Collaborateur getCurrentUser() {
+    /**
+     * Récupère l'utilisateur actuellement connecté (si un utilisateur est connecté).
+     * @return une entité Collaborateur
+     */
+    public Collaborateur getCurrentUser() throws EntityNotFoundException {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
         if (!(authentication instanceof AnonymousAuthenticationToken)) {
             String username = authentication.getName();
             return this.findByPseudo(username);
         }
-        return null;
+
+        throw new NotAuthenticatedException("Pas d'utilisateur connecté");
     }
 
+    /**
+     * Méthode utilisée par /login géré par Spring Security pour l'authentification.
+     * Retourne une entité UserDetails.
+     */
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Collaborateur collaborateur = collaborateurRepository.findByPseudo(username);
+        Optional<Collaborateur> optionalCollaborateur = collaborateurRepository.findByPseudo(username);
 
-        if (collaborateur == null) {
+        // Si aucun collaborateur trouvé à partir du pseudo
+        if (optionalCollaborateur.isEmpty()) {
             throw new UsernameNotFoundException("Collaborateur inconnu");
         }
+
+        Collaborateur collaborateur = optionalCollaborateur.get();
 
         // Déterminer le rôle en fonction du champ "admin"
         String role = collaborateur.getAdmin() ? "ROLE_ADMIN" : "ROLE_USER";
@@ -60,13 +88,22 @@ public class CollaborateurService implements UserDetailsService {
         return new User(collaborateur.getPseudo(), collaborateur.getPassword(), authorities);
     }
 
-
-    public List<Collaborateur> getAllCollaborateurs() {
-        return collaborateurRepository.findAll();
+    /**
+     * Retourne un set de collaborateur
+     * @return
+     */
+    public Set<Collaborateur> getAllCollaborateurs() {
+        return collaborateurRepository.findAllCollaborateurs();
     }
 
-    public Optional<Collaborateur> getCollaborateurById(int id) {
-        return collaborateurRepository.findById(id);
+    public Collaborateur getCollaborateurById(int id) throws EntityNotFoundException {
+        Optional<Collaborateur> collaborateur = collaborateurRepository.findById(id);
+
+        if (collaborateur.isEmpty()) {
+            throw new EntityNotFoundException("Pas de collaborateur trouvé à l'ID " + id);
+        }
+
+        return collaborateur.get();
     }
 
     public Collaborateur createCollaborateur(Collaborateur collaborateur) {
@@ -77,15 +114,15 @@ public class CollaborateurService implements UserDetailsService {
         return collaborateurRepository.save(collaborateur);
     }
 
-    public void deleteCollaborateur(int id) {
-        collaborateurRepository.deleteById(id);
+    public void deleteCollaborateur(int id) throws EntityNotFoundException, RuntimeException {
+        if (collaborateurRepository.existsById(id)) {
+            collaborateurRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Pas de collaborateur trouvé pour l'ID " + id);
+        }
     }
 
     public Collaborateur saveCollaborateur(Collaborateur collaborateur) {
         return collaborateurRepository.save(collaborateur);
-    }
-
-    public Optional<Collaborateur> login(String pseudo, String password) {
-        return collaborateurRepository.findByPseudoAndPassword(pseudo, password);
     }
 }
